space optimisation of print()
Removed need for print(float f) by casting state info to int - saves
about 500 bytes

+++ software/firmware/LogoBotWifiScribbler/LogoBotWifiScribbler.ino
@@ -95,11 +95,11 @@ static void handleCollision(byte collisionData)
 static void showStatus()
 {
   // Format: "X Y ang qSize"
-  Serial.print(bot.state.x);
+  Serial.print((int)bot.state.x);
   Serial.print(" ");
-  Serial.print(bot.state.y);
+  Serial.print((int)bot.state.y);
   Serial.print(" ");
-  Serial.print(bot.state.ang);
+  Serial.print((int)bot.state.ang);
   Serial.print(" ");
   Serial.println(cmdQ.pending());
 }
